# Performance improvements in/near Molecule API and RDKit wrapper